% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_5.14 copy.R
\name{clinicalmodel_linear}
\alias{clinicalmodel_linear}
\title{Linear clinical model}
\usage{
clinicalmodel_linear(GBM_data, nruns, delta, take_log)
}
\arguments{
\item{GBM_data}{fitted mechanistic model}

\item{nruns}{nruns}

\item{delta}{delta}

\item{take_log}{TRUE/FALSE of taking logrithm of clinical outcome}
}
\description{
Linear clinical model
}
