% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_5.14 copy.R
\name{clinicalmodel_survbart}
\alias{clinicalmodel_survbart}
\title{BART clinical model}
\usage{
clinicalmodel_survbart(GBM_data, nruns, delta)
}
\arguments{
\item{GBM_data}{fitted mechanistic model}

\item{nruns}{nruns}

\item{delta}{delta}
}
\description{
BART clinical model
}
